+++ examples/data-sources/coolify_s3_storage/data-source.tf
@@ -1,0 +1,3 @@
+data "coolify_s3_storage" "example" {
+  uuid = "existing-s3-storage-uuid"
+}