+++ examples/data-sources/coolify_s3_storages/data-source.tf
@@ -1,0 +1,5 @@
+data "coolify_s3_storages" "all" {}
+
+output "storage_names" {
+  value = [for s in data.coolify_s3_storages.all.storages : s.name]
+}